rename external folder
