+++ external_/connections_external_connection_item_request_builder.go
