+++ external_/connections_item_groups_count_request_builder.go
