+++ external_/connections_item_groups_request_builder.go
