+++ external_/connections_item_items_count_request_builder.go
