+++ external_/connections_item_items_external_item_item_request_builder.go
