+++ external_/connections_item_items_item_activities_count_request_builder.go
