+++ external_/connections_item_items_item_add_activities_response.go
