+++ external_/connections_item_operations_connection_operation_item_request_builder.go
