+++ external_/connections_item_operations_count_request_builder.go
