+++ external_/connections_item_operations_request_builder.go
