+++ external_/connections_item_schema_request_builder.go
